requirements.txt
flask==3.0.3
scikit-learn==1.5.1
pandas==2.2.2
numpy==2.1.1
python-whois==0.8.0
joblib==1.4.2
httpx==0.27.0
tldextract==5.1.2